test byte serialize

+++ ListNodeSerializer/Nodes/NodeInfo.cs
@@ -2,8 +2,8 @@
 
 internal class NodeInfo
 {
-    public long Order { get; set; }
-    public long Id { get; set; }
-    public long? RandId { get; set; }
+    public int Order { get; set; }
+    public int Id { get; set; }
+    public int? RandId { get; set; }
     public string Data { get; set; }
 }
